Fix deleteat! on ChainedVector when multiple chunks are deleted (#39)
* Fix deleteat! on ChainedVector when multiple chunks are deleted

Fixes #38. The issue here was when deleting spanned multiple chunks of a
ChainedVector, we were then going through and "cleaning up" the chunks
after filtering and were deleting the chunks via forward iteration,
which messed up deleting as the later chunks changed index. By deleting
the chunks in reverse, we ensure hte correct chunks are removed.

* fix ambiguity

+++ src/SentinelArrays.jl
@@ -227,7 +227,7 @@ Base.@propagate_inbounds function Base.setindex!(A::SentinelArray{T, N, S, V}, v
 end
 
 # other AbstractArray functions
-function Base.reverse(A::SentinelVector, s=first(LinearIndices(A)), n=last(LinearIndices(A)))
+function Base.reverse(A::SentinelVector, s::Integer=first(LinearIndices(A)), n::Integer=last(LinearIndices(A)))
     return SentinelArray(reverse(parent(A), s, n), A.sentinel, A.value)
 end
